requirements.txt
fastapi
uvicorn[standard]
pydantic
python-multipart

python-socketio

PyJWT
pyjwt[crypto]
bcrypt
passlib[bcrypt]
uuid

pytest

requests
pymongo